refactor(vm): Move Opaque to its own module
